guardian-coordinator/requirements.txt
# Guardian Coordinator Dependencies
# Python 3.11+ required

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
python-multipart==0.0.18
python-dotenv==1.0.0

# Discord bot
discord.py==2.3.2
aiohttp==3.9.1

# Database
supabase==2.3.4
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
click==8.1.7
colorama==0.4.6
